Add routers, refactor code
